Add joint command subscribers with debug logging
- Added harmony_topics.cpp/h for ROS topic subscribers
- Implemented subscribers for torque, stiffness, and position commands
- Added PLOGI logging to show which arm and fields are being updated
- Added PLOGD logging to show preserved values when updating single field types
- Logging helps track override updates and verify value preservation

+++ tools/data_exerciser.cpp
@@ -306,7 +306,7 @@ int main(int argc, const char** argv) {
 
     for (int i=0; i<=nSteps; i++) {
         robotStartPosition = getCurrentArmPositionsAsDataLine(&info);
-        auto overrides = data2override(robotStartPosition, i/nSteps);
+        auto overrides = data2override(robotStartPosition, static_cast<double>(i)/static_cast<double>(nSteps));
 
         left->setJointsOverride(overrides.leftOverrides);
         right->setJointsOverride(overrides.rightOverrides);
